Update user identity fetching with cache
